Rename package cert-generator with certgenerator
