+++ pkg/certgenerator/v1beta1/generator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta1
+package certgenerator
 
 import (
 	"bytes"
